+++ test/performance/runtests_benchmark.jl
@@ -1,21 +1,11 @@
 using Jute
 
-run_options = RunOptions(Dict(
-    :dont_add_runtests_path => true,
-    :test_module_prefix => "",
-    :test_file_postfix => ".test.jl",
-    :include_only => nothing,
-    :exclude => nothing,
-    :include_only_tags => [],
-    :exclude_tags => [],
-    :max_fails => 0,
-    :capture_output => false,
-    :verbosity => 0))
-
 runtests_dir = Jute.get_runtests_dir()
 test_files = [joinpath(runtests_dir, "runtests_benchmark_testcases.jl")]
 
 test_include_only = (ARGS[1] == "test_include_only")
+
+run_options = build_run_options(options=Dict(:verbosity => 0))
 
 tic()
 obj_dict = Jute.include_test_files!(test_files)